# specs/001-job-matching-system/contracts/detailed-api-spec.yaml
openapi: 3.0.3
info:
  title: Job Matching System - Detailed API Specification
  description: |
    バイト求人マッチングシステムの詳細API仕様書

    ## システム概要
    - 10万件の求人データから1万人のユーザーに最適な40件を毎日選定
    - パーソナライズされたメール配信準備
    - 3段階スコアリング（基礎・SEO・パーソナライズ）
    - 6セクション構成のメール生成

    ## 技術スタック
    - Backend: Python 3.11 + FastAPI
    - Database: Supabase (PostgreSQL)
    - AI: OpenAI GPT-4/5 nano
    - Monitoring: Real-time SQL dashboard
  version: 2.0.0
  contact:
    name: Job Matching System Team
    email: dev@job-matching.com
    url: https://github.com/job-matching/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v2
    description: Development server
  - url: https://staging-api.job-matching.com/v2
    description: Staging server
  - url: https://api.job-matching.com/v2
    description: Production server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: ユーザー一覧取得
      description: |
        ユーザー一覧を取得。
        フィルタリング、ページネーション対応。
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, deleted]
            default: active
        - name: location
          in: query
          schema:
            type: string
          description: 居住地フィルタ
        - name: age_min
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: age_max
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: registered_after
          in: query
          schema:
            type: string
            format: date
        - name: registered_before
          in: query
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, last_login, engagement_score]
            default: updated_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: ユーザー一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Users]
      summary: ユーザー作成
      description: 新しいユーザーを作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: ユーザーが既に存在
  /users/{user_id}:
    get:
      tags: [Users]
      summary: ユーザー詳細取得
      description: 指定されたIDのユーザー詳細を取得
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ユーザー詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Users]
      summary: ユーザー更新
      description: 指定されたIDのユーザーを更新
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Users]
      summary: ユーザー削除
      description: 指定されたIDのユーザーを削除（論理削除）
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: ユーザー削除成功
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{user_id}/profile:
    get:
      tags: [Users]
      summary: ユーザープロファイル取得
      description: ユーザーの詳細プロファイル情報を取得
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ユーザープロファイル
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Users]
      summary: ユーザープロファイル更新
      description: ユーザーのプロファイル情報を更新
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: プロファイル更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/{user_id}/preferences:
    get:
      tags: [Users]
      summary: ユーザー設定取得
      description: ユーザーの設定・環境設定を取得
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ユーザー設定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Users]
      summary: ユーザー設定更新
      description: ユーザーの設定を更新
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdateRequest'
      responses:
        '200':
          description: 設定更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        error_code:
          type: string
          description: エラーコード
        details:
          type: array
          items:
            type: string
          description: 詳細エラー情報
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
          description: トレースID

    # ===== 求人関連スキーマ =====
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        age:
          type: integer
          minimum: 0
          maximum: 100
        gender:
          type: string
          enum: [male, female, other, not_specified]
        location:
          type: object
          properties:
            prefecture:
              type: string
            city:
              type: string
            address:
              type: string
        status:
          type: string
          enum: [active, inactive, suspended, deleted]
        email_verified:
          type: boolean
        last_login:
          type: string
          format: date-time
        engagement_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - status
    UserCreateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 100
        age:
          type: integer
          minimum: 0
          maximum: 100
        gender:
          type: string
          enum: [male, female, other, not_specified]
        location:
          type: object
          properties:
            prefecture:
              type: string
            city:
              type: string
            address:
              type: string
      required:
        - email
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        age:
          type: integer
          minimum: 0
          maximum: 100
        gender:
          type: string
          enum: [male, female, other, not_specified]
        location:
          type: object
          properties:
            prefecture:
              type: string
            city:
              type: string
            address:
              type: string
        status:
          type: string
          enum: [active, inactive, suspended]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    UserProfile:
      type: object
      properties:
        user_id:
          type: integer
        basic_info:
          $ref: '#/components/schemas/User'
        work_preferences:
          type: object
          properties:
            preferred_occupations:
              type: array
              items:
                type: string
            preferred_locations:
              type: array
              items:
                type: string
            salary_expectation:
              type: object
              properties:
                min:
                  type: integer
                max:
                  type: integer
            work_style:
              type: array
              items:
                type: string
                enum: [part_time, full_time, short_term, long_term,
                       weekend_only, night_shift, flexible]
        experience:
          type: object
          properties:
            work_history:
              type: array
              items:
                type: object
                properties:
                  occupation:
                    type: string
                  duration_months:
                    type: integer
                  skills:
                    type: array
                    items:
                      type: string
            education:
              type: string
              enum: [high_school, vocational, university, graduate, other]
        personal_info:
          type: object
          properties:
            transportation:
              type: array
              items:
                type: string
                enum: [walk, bicycle, public_transport, car, motorcycle]
            availability:
              type: object
              properties:
                weekdays:
                  type: boolean
                weekends:
                  type: boolean
                mornings:
                  type: boolean
                afternoons:
                  type: boolean
                evenings:
                  type: boolean
                nights:
                  type: boolean
        updated_at:
          type: string
          format: date-time
    UserProfileUpdateRequest:
      type: object
      properties:
        work_preferences:
          type: object
          properties:
            preferred_occupations:
              type: array
              items:
                type: string
            preferred_locations:
              type: array
              items:
                type: string
            salary_expectation:
              type: object
              properties:
                min:
                  type: integer
                max:
                  type: integer
            work_style:
              type: array
              items:
                type: string
        experience:
          type: object
          properties:
            work_history:
              type: array
              items:
                type: object
                properties:
                  occupation:
                    type: string
                  duration_months:
                    type: integer
                  skills:
                    type: array
                    items:
                      type: string
            education:
              type: string
              enum: [high_school, vocational, university, graduate, other]
        personal_info:
          type: object
          properties:
            transportation:
              type: array
              items:
                type: string
            availability:
              type: object
...
    UserPreferences:
      type: object
      properties:
        user_id:
          type: integer
        notification_settings:
          type: object
          properties:
            email_enabled:
              type: boolean
            email_frequency:
              type: string
              enum: [daily, weekly, bi_weekly, monthly]
            sms_enabled:
              type: boolean
            push_enabled:
              type: boolean
        privacy_settings:
          type: object
          properties:
            profile_visibility:
              type: string
              enum: [public, private, employers_only]
            search_tracking:
              type: boolean
            behavior_tracking:
              type: boolean
        recommendation_settings:
          type: object
          properties:
            algorithm_preference:
              type: string
              enum: [collaborative, content_based, hybrid]
            diversity_weight:
              type: number
              format: float
              minimum: 0
              maximum: 1
            freshness_weight:
              type: number
              format: float
              minimum: 0
              maximum: 1
        updated_at:
          type: string
          format: date-time
    UserPreferencesUpdateRequest:
      type: object
      properties:
        notification_settings:
          type: object
          properties:
            email_enabled:
              type: boolean
            email_frequency:
              type: string
              enum: [daily, weekly, bi_weekly, monthly]
            sms_enabled:
              type: boolean
            push_enabled:
              type: boolean
        privacy_settings:
          type: object
          properties:
            profile_visibility:
              type: string
              enum: [public, private, employers_only]
            search_tracking:
              type: boolean
            behavior_tracking:
              type: boolean
        recommendation_settings:
          type: object
          properties:
            algorithm_preference:
              type: string
              enum: [collaborative, content_based, hybrid]
            diversity_weight:
              type: number
              format: float
              minimum: 0
              maximum: 1
            freshness_weight:
              type: number
              format: float
              minimum: 0
              maximum: 1
...
  responses:
    BadRequest:
      description: リクエストが無効です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        APIキー認証
        ヘッダーに `X-API-Key: your-api-key` を含めてください

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT認証
        ヘッダーに `Authorization: Bearer <token>` を含めてください
security:
  - ApiKeyAuth: []
  - BearerAuth: []

# ===========================================
# 追加情報・拡張仕様
# ===========================================
...